Refactor movie management to support redesigned model
Updated movie-related logic to utilize `PeliculaRedesign` class. Enhanced the GUI, added a new panel for adding movies, and adjusted DAO methods for consistent handling of the refined model structure.

+++ src/main/java/models/PeliculaRedesign.java
@@ -8,7 +8,7 @@ public class PeliculaRedesign {
     private int id;
     private String titulo;
     private String director;
-    private double duracion;
+    private int duracion;
     private double precio;
     private enum categoria{
         COMEDIA,
@@ -18,7 +18,7 @@ public class PeliculaRedesign {
         AVENTURA
     };
 
-    public PeliculaRedesign(int id, String titulo, String director, double duracion, double precio) {
+    public PeliculaRedesign(int id, String titulo, String director, int duracion, double precio) {
         if (Validate.isInteger(String.valueOf(id)) &&
                 Validate.isString(titulo) &&
                 Validate.isString(director) &&
